Project 2 - Custom Copilot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aniso8601==10.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argcomplete==3.6.3
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
azure-ai-ml==1.30.0
azure-common==1.1.28
azure-core==1.36.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==4.14.2
azure-identity==1.25.1
azure-mgmt-core==1.6.0
azure-search-documents==11.4.0
azure-monitor-opentelemetry==1.8.2
azure-monitor-opentelemetry-exporter==1.0.0b45
azure-search-documents==11.6.0
azure-storage-blob==12.27.1
azure-storage-file-datalake==12.22.0
azure-storage-file-share==12.23.1
backports.tarfile==1.2.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
fastapi==0.121.2
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
flask-restx==1.3.2
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-search-results==2.4.1
googleapis-common-protos==1.72.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==24.3.1
keyrings.alt==5.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.0
numpy==2.0.2
oauthlib==3.3.1
openai==2.8.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-dbapi==0.59b0
opentelemetry-instrumentation-django==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-flask==0.59b0
opentelemetry-instrumentation-psycopg2==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-urllib==0.59b0
opentelemetry-instrumentation-urllib3==0.59b0
opentelemetry-instrumentation-wsgi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
packaging==25.0
pandas==2.3.3
pillow==11.0.0
promptflow==1.18.1
promptflow-azure==1.18.1
promptflow-core==1.18.1
promptflow-devkit==1.18.1
promptflow-tools==1.6.1
promptflow-tracing==1.18.1
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydash==7.0.7
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
SecretStorage==3.3.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
strictyaml==1.7.3
tabulate==0.9.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0